sanchai-weather-backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-openai
